- Effect Factory - Removal of other effects factories - Addition of MapDataJsonParser - Hot Fixes
