--- src/main/java/edu/fiuba/algo3/modelo/factories/BeastFactory.java
@@ -1,11 +1,0 @@
-package edu.fiuba.algo3.modelo.factories;
-
-import edu.fiuba.algo3.modelo.factories.EffectFactory;
-import edu.fiuba.algo3.modelo.squares.Beast;
-import edu.fiuba.algo3.modelo.squares.Effect;
-
-public class BeastFactory extends EffectFactory {
-    public Effect createEffect(){
-        return new Beast();
-    }
-}
