--- src/main/java/edu/fiuba/algo3/modelo/factories/NullEffectFactory.java
@@ -1,11 +1,0 @@
-package edu.fiuba.algo3.modelo.factories;
-
-import edu.fiuba.algo3.modelo.factories.EffectFactory;
-import edu.fiuba.algo3.modelo.squares.Effect;
-import edu.fiuba.algo3.modelo.squares.NullEffect;
-
-public class NullEffectFactory extends EffectFactory {
-    public Effect createEffect(){
-        return new NullEffect();
-    }
-}
