+++ src/main/java/edu/fiuba/algo3/modelo/mapJsonParser/ObstaclesRewardsParser.java
@@ -5,8 +5,6 @@ import java.io.FileReader;
 import java.io.IOException;
 import java.util.ArrayList;
 
-import edu.fiuba.algo3.modelo.RandomResult.DiceFactory;
-import edu.fiuba.algo3.modelo.RandomResult.RandomResult;
 import edu.fiuba.algo3.modelo.squares.*;
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
@@ -80,5 +78,4 @@ public class ObstaclesRewardsParser {
             throw new MapFileCouldNotBeParsed();
         }
     }
-
 }
