+++ src/main/test/edu/fiuba/algo3/entrega_2/UseCase17.java
@@ -8,6 +8,7 @@ import edu.fiuba.algo3.modelo.mapJsonParser.MapFileCouldNotBeParsed;
 import edu.fiuba.algo3.modelo.mapJsonParser.MapFileFailedToOpenOrClose;
 import edu.fiuba.algo3.modelo.mapJsonParser.MapFileNotFound;
 import edu.fiuba.algo3.modelo.squares.Position;
+import edu.fiuba.algo3.modelo.squares.Square;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -16,21 +17,19 @@ import java.util.ArrayList;
 public class UseCase17 {
     @Test
     public void test17AMapIsBuiltCorrectlyBasedOnJSONFile() throws MapFileNotFound, MapFileFailedToOpenOrClose, MapFileCouldNotBeParsed, InvalidMapFile {
-        Map map;
-        ArrayList<Position> path;
+        ArrayList<Square> path;
         MapFacade facade = new MapFacade();
         Gladiator aGladiator = new Gladiator();
         int expectedInitialEnergy = 0;
         int expectedSecondEnergy = 20;
         int expectedThirdEnergy = 35;
 
-        map = facade.loadMap();
-        path = map.getPath();
+        path = facade.loadMap();
         assertEquals(expectedInitialEnergy, aGladiator.getEnergy());
-        path.get(0).receivePiece(aGladiator);
+        path.get(0).affect(aGladiator);
         assertEquals(expectedSecondEnergy, aGladiator.getEnergy());
 
-        path.get(2).receivePiece(aGladiator);
+        path.get(2).affect(aGladiator);
 
         assertEquals(expectedThirdEnergy, aGladiator.getEnergy());
     }
